Fix ServiceWorker Problems
We had serious problems with the ServiceWorker of our old
website loading the old page instead of the new one. This commit fixes
this problem by shipping a service-worker.js file that cleans the cache.
After one more page reload the new page gets served. In addition the log
messages are prettified when trying to unregister the old ServiceWorker.